Punkte mit Komma-Werten, Einsatz/Dienstbuch/Dienststunden-Punkteregeln, Dienstbücher-Liste wie Einsätze
- person_punkte.punkte ist jetzt Float (Komma-Punkte in der DB erlaubt),
  angezeigt wird der Person immer nur die gerundete Gesamtsumme.
- Neue Punkte-Regeln: Einsatz-Abschluss und Dienstbuch-Abschluss vergeben je
  Teilnahme konfigurierbare Punkte; werden bei Wiedereröffnung automatisch
  zurückgenommen und bei erneutem Abschließen/Schließen erneut vergeben.
- Dienstbücher können jetzt auch manuell geschlossen/wieder geöffnet werden
  (neue Endpunkte, vorher nur über den Auto-Schließen-Job möglich).
- Dienststunden-Punkte sind jetzt "Punkte je Stunde", präzise auf die Minute
  berechnet (z. B. 1,5 Std. = 1,5-facher Wert) statt fix pro Eintrag.
- Punkte-Einstellungsseite um die drei neuen Regeln erweitert, Dezimal-Eingabe
  möglich.
- Listen > Dienstbücher: Status-Spalte/-Filter sowie Schließen/Wieder-öffnen-
  Aktionen ergänzt, analog zur Einsätze-Liste.

+++ backend/app/models/person_punkt.py
@@ -1,6 +1,6 @@
 from datetime import date, datetime
 
-from sqlalchemy import Date, DateTime, ForeignKey, Integer, String, func
+from sqlalchemy import Date, DateTime, Float, ForeignKey, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -17,7 +17,10 @@ class PersonPunkt(Base):
     person_id: Mapped[int] = mapped_column(
         ForeignKey("personen.id", ondelete="CASCADE"), nullable=False
     )
-    punkte: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Kommazahlen erlaubt (z. B. Dienststunden-Punkte je Stunde, präzise auf
+    # die Minute berechnet) – angezeigt wird der Person immer nur die
+    # gerundete Gesamtsumme (siehe stammdaten_service.gesamtpunkte_batch).
+    punkte: Mapped[float] = mapped_column(Float, nullable=False)
     grund: Mapped[str] = mapped_column(String(64), nullable=False)
     gueltig_bis: Mapped[date] = mapped_column(Date, nullable=False)
     # "halten": Punkte bleiben bis gueltig_bis voll erhalten (Standard).
